pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nian-kantoku"
version = "0.1.0"
description = "FMVP anime storyboard-to-video generator"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "PyYAML>=6.0",
  "requests>=2.31.0",
  "rich>=13.7.0",
  "volcengine-python-sdk[ark]>=5.0.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
nian-kantoku = "nian_kantoku.interface.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
